resmgr: delay NRI plugin setup after configuration.
Delay setting up the NRI plugin interface after having
acquired the startup configuration, or more precisely,
after having queryied the upstream agent interface for
initial configuration data.

This initial query has a timeout of 1 second. If the
agent is not around the query always times out taking
the full 1 second even if there is no agent socket in
place.

If we instantiate the NRI plugin/stub before the query
but start the stub only after and the agent is down, or
slow, the plugin registratition sequence will time out.
The timeout is measure between plugin connection (stub
creation) and plugin registration (stub Start()/Run())
(and it is hardcoded to 1 second as of this writing).