e2e: Launch process and wait it to start properly
When launching the process, wait until it starts properly before
continuing the script. Earlier there was fixed sleeps after launching
process like cri-resmgr, but this is not good if govm is running
inside a slow VM. It might happen that process starts slow because
of machine load and large binary size and the process is not yet
fully running when we check the process status.